requirements.txt
joblib>=1.0.0
numpy>=1.19.0
pandas>=1.1.0
python-igraph>=0.9.0
scikit-learn>=0.24.0
scipy>=1.6.0
tqdm>=4.40.0a